# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "xinzhi-daoxue-api",
    "description": "芯智导学阶段 2.2 API。统一 Agent 注册、快速路由、受控降级与多星辰工作流共用同一 Provider 调用链。",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/debug/rag/run": {
      "post": {
        "tags": [
          "rag-debug"
        ],
        "summary": "Run Debug",
        "operationId": "run_debug_api_v1_debug_rag_run_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DebugRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Run Debug Api V1 Debug Rag Run Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/debug/rag/prewarm": {
      "post": {
        "tags": [
          "rag-debug"
        ],
        "summary": "Prewarm",
        "description": "Explicitly load selected local models without loading them at startup.",
        "operationId": "prewarm_api_v1_debug_rag_prewarm_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PrewarmRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Prewarm Api V1 Debug Rag Prewarm Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/debug/rag/compare": {
      "post": {
        "tags": [
          "rag-debug"
        ],
        "summary": "Compare",
        "operationId": "compare_api_v1_debug_rag_compare_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompareRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Compare Api V1 Debug Rag Compare Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/debug/rag/traces/{trace_id}": {
      "get": {
        "tags": [
          "rag-debug"
        ],
        "summary": "Trace",
        "operationId": "trace_api_v1_debug_rag_traces__trace_id__get",
        "parameters": [
          {
            "name": "trace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Trace Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Trace Api V1 Debug Rag Traces  Trace Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/debug/rag/eval": {
      "post": {
        "tags": [
          "rag-debug"
        ],
        "summary": "Evaluate",
        "operationId": "evaluate_api_v1_debug_rag_eval_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Evaluate Api V1 Debug Rag Eval Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CompareRequest": {
        "properties": {
          "question": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Question"
          },
          "course_id": {
            "type": "string",
            "enum": [
              "CT",
              "AE",
              "DE"
            ],
            "title": "Course Id",
            "default": "CT"
          },
          "intent": {
            "type": "string",
            "enum": [
              "general_qa",
              "explain_concept",
              "summarize_knowledge",
              "learning_advice",
              "solve_problem"
            ],
            "title": "Intent",
            "default": "explain_concept"
          },
          "response_depth": {
            "type": "string",
            "enum": [
              "brief",
              "standard",
              "deep"
            ],
            "title": "Response Depth",
            "default": "standard"
          },
          "conversation_summary": {
            "type": "string",
            "maxLength": 1000,
            "title": "Conversation Summary",
            "default": ""
          },
          "previous_answer_summary": {
            "type": "string",
            "maxLength": 1000,
            "title": "Previous Answer Summary",
            "default": ""
          },
          "use_rag": {
            "type": "boolean",
            "title": "Use Rag",
            "default": true
          },
          "include_images": {
            "type": "boolean",
            "title": "Include Images",
            "default": false
          },
          "use_reranker": {
            "type": "boolean",
            "title": "Use Reranker",
            "default": false
          },
          "allow_cloud": {
            "type": "boolean",
            "title": "Allow Cloud",
            "default": false
          },
          "request_id": {
            "type": "string",
            "maxLength": 128,
            "title": "Request Id",
            "default": ""
          },
          "comparison_mode": {
            "type": "string",
            "enum": [
              "rag_vs_no_rag",
              "cloud_vs_local"
            ],
            "title": "Comparison Mode",
            "default": "rag_vs_no_rag"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "question"
        ],
        "title": "CompareRequest"
      },
      "DebugRunRequest": {
        "properties": {
          "question": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Question"
          },
          "course_id": {
            "type": "string",
            "enum": [
              "CT",
              "AE",
              "DE"
            ],
            "title": "Course Id",
            "default": "CT"
          },
          "intent": {
            "type": "string",
            "enum": [
              "general_qa",
              "explain_concept",
              "summarize_knowledge",
              "learning_advice",
              "solve_problem"
            ],
            "title": "Intent",
            "default": "explain_concept"
          },
          "response_depth": {
            "type": "string",
            "enum": [
              "brief",
              "standard",
              "deep"
            ],
            "title": "Response Depth",
            "default": "standard"
          },
          "conversation_summary": {
            "type": "string",
            "maxLength": 1000,
            "title": "Conversation Summary",
            "default": ""
          },
          "previous_answer_summary": {
            "type": "string",
            "maxLength": 1000,
            "title": "Previous Answer Summary",
            "default": ""
          },
          "use_rag": {
            "type": "boolean",
            "title": "Use Rag",
            "default": true
          },
          "include_images": {
            "type": "boolean",
            "title": "Include Images",
            "default": false
          },
          "use_reranker": {
            "type": "boolean",
            "title": "Use Reranker",
            "default": false
          },
          "allow_cloud": {
            "type": "boolean",
            "title": "Allow Cloud",
            "default": false
          },
          "request_id": {
            "type": "string",
            "maxLength": 128,
            "title": "Request Id",
            "default": ""
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "question"
        ],
        "title": "DebugRunRequest"
      },
      "EvalRequest": {
        "properties": {
          "group": {
            "type": "string",
            "enum": [
              "all",
              "CT",
              "AE",
              "DE",
              "boundary",
              "degradation"
            ],
            "title": "Group",
            "default": "all"
          },
          "allow_cloud": {
            "type": "boolean",
            "title": "Allow Cloud",
            "default": false
          },
          "limit": {
            "type": "integer",
            "maximum": 60.0,
            "minimum": 1.0,
            "title": "Limit",
            "default": 60
          }
        },
        "type": "object",
        "title": "EvalRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PrewarmRequest": {
        "properties": {
          "models": {
            "items": {
              "type": "string",
              "enum": [
                "text",
                "image",
                "reranker"
              ]
            },
            "type": "array",
            "title": "Models"
          }
        },
        "type": "object",
        "title": "PrewarmRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}